pyproject.toml
[tool.poetry]
name = "Online_cinema"
version = "0.1.0"
description = "Your project description"
authors = ["Maxim <maximuschampion2002@gmail.com>"]
readme = "README.md"

packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = ">=2.7.0,<3.0"
pydantic-settings = "^2.10"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
sqlalchemy = "^2.0.0"
asyncpg = "^0.27.0"
celery = "^5.3.0"
redis = "^4.5.1"
python-multipart = "^0.0.6"
email-validator = "^2.0"
aiofiles = "^23.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.0"
aiosmtplib = "^1.1.6"
python-jose = "^3.3"
aioboto3 = "^13.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.10"



[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.20.3"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"
alembic = "^1.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
